CSSMin: Remove duplicate code in remap() and fix Phan warning
The two code blocks for remap-only and remap-with-maybe-embed
are identical now that the various code for IE 6 and IE 8 has
been removed (I111c456cfc3e48b).

These code blocks only differed in the variable name they assign
to, and the `$embedData` parameter which was hardcoded to false
in the first half, and set based on the parameter/match in the
second half. Given that both halves were unconditionally executed
in prod, this was also inefficient and rather wasteful. I've
kept the second version only, documented it, and added a test for
the logic in question, which was also absnet.

Folding these together also removes the need for the clever return
value which previously triggered the "PhanPossiblyUndeclaredVariable"
code smell warning.

Change-Id: I7cafe545ca195139a88f0b062f5f5486ab437e00