CSSMin: Remove old IE 8 compat code and warn for too-large embeds
We no longer supports IE 8 for Grade C, per T248061 and
<https://www.mediawiki.org/wiki/Compatibility>. As such, producing
data URIs longer than what IE 8 could handle is no longer a concern.

However, we have significantly reduced use of `@embed` in the last
two years (see T121730, T118514, and T120984), so for now I'm keeping
the limit in-place with a run time warning.

It generally makes sense not to embed large files. Based on what we
see in the wild, we may tweak the limit a bit, but as a general concept
it makes sense to preserve long-term, even if we end up raise
it slightly.

Change-Id: I5cc4761f2fcf03032a43e9329ded2ecede52dd63