JavaScriptMinifier: Apply true/false minification in more situations
The code that changes true to !0 and false to !1 was only applied in the
EXPRESSION and PROPERTY_EXPRESSION states, but not in the
EXPRESSION_NO_NL, EXPRESSION_TERNARY and PAREN_EXPRESSION states, which
are very similar. Add an array of EXPRESSION-like states, and apply
true/false minification in all of them.

Change-Id: I9bf725cb842f3e0681b8036c3d3175365887805f