JavaScriptMinifier: Consolidate *_FUNC states into a single FUNC state
Instead of having separate EXPRESSION_FUNC, EXPRESSION_TERNARY_FUNC,
PROPERTY_EXPRESSION_FUNC and PAREN_EXPRESSION_FUNC states, use one
single FUNC state. This is possible because the only difference is which
state to go to after the function's closing } , so instead of encoding
that difference in different state names, we push that value to the
stack before moving to the FUNC state.

This is just a simplification of the state machine, it shouldn't affect
the output at all.

Change-Id: Ibeb332af0760cd16809473b308443e42e1ac5a37