CSSMin: Fix remapping of path-only URL when base dir is server-less root
A similar issue was previously fixed in MediaWiki with commit
I16abf482e3a against ResourceLoaderFileModule. There, the issue has
to do with the remote path, which is then later expanded to include
a server so long as it starts with "/". That commit did not involve
changing CSSMin, because by the time it reached this code, it had
been turned into `https://server.example/`.

What we did not yet test for is if a bare "/" makes it into CSSMin
as remote base path. While this is not possible for file-based
modules, it is possible in MediaWiki for wiki-based modules where
URLs are remapped against the current domain without needing to
hardcode it again. The problem was that we stripped off the "/" here
and thus made it invalid for use as base URL.

Bug: T282280
Change-Id: I4a806a688c69fe8525271d3a55a23be4f857e89f