JavaScriptMinifer: Recognize `...` as a single token
The code for recognizing punctuation tokens (which can be of variable
length) assumed that, for every multi-character punctuation token,
every prefix of that token was also a valid token. For example, for
!== it assumed that ! and != are valid tokens (which is true). This
assumption held for every punctuation token except for ... (because ..
is not a valid token).

This caused the state machine to see ... as three DOT tokens rather than
a single UN_OP token. That didn't cause any state machine issues
(because it would only have made a difference if ... was followed by a
keyword like if or function, which isn't legal), but it did make it
possible for a ... token to be split by newline insertion.

Fix this by dropping the prefix assumption, and instead checking all
possible token lengths from long to short.

Bug: T287526
Change-Id: I451eae5f9e6f07305de6c28d0caf6f5903886e7f