JavaScriptMinifier: Remove `true/!0` `false/!1` transformations
There seem to be too many edge cases where this doesn't work right.
We've fixed some of them over the years, but there are still some
left and it's not clear to me that the gains are worth it in the face
of on-the-wire compression (gzip), and the would-be cost increase of
adding additional parentheses or excluding it with additional runtime
complexity. There is also the cost of additional token parsing and
runtime operations on the client end, which seems to be increasingly
where the bottleneck lies.

We may introduce this in a more limited form in the future as part of
an effort to incorporate a number of narrow save and clear-win
transformations, but for now I'd like to just resolve a known bug
and move on to other and bigger things:

```
$ echo 'true.toString()' | bin/minify js
true.toString() # OK

$ echo 'x = true.toString()' | bin/minify js
x=!0.toString() # SyntaxError
```

Bug: T237042
Change-Id: If10c7c5961a51f9f3ffb801ab6c40e55d9d65b0b