JavaScriptMinifier: Correctly recognize \` in template strings
Backticks were always considered to end a template string, even if they
were escaped. For example, in `foo\`bar`, the literal was incorrectly
considered to end after `foo\`, and bar` was considered code.

This was caused by a very subtle bug in the loop that processes template
strings. There was even a test case that asserted the incorrect
behavior, probably because of confusion over the need to double-escape a
backslash.

Bug: T296058
Change-Id: I205fb720b7f2f4aa041fbb8c5360cfa78178e3cd