JavaScriptMinifier: Fix handling of property on dotless number literal
The dot punctuation is in $opChars which we generally treat as not
needing whitespace with surrounding tokens that aren't in $opChars.
This means `a + b` can become `a+b` and while vastly different in
meaning in JS, we apply the same to `a . bar`.

This holds up for many kinds of numbers as well, such as `42.0 .bar`
can become `42.0.bar`. 42. .bar`, and with implied factions also,
`42. .bar` can become `42..bar`.

However, in the case of a dotless number literal, the only way that
we can find a dot as its own token is if it was separated by
whitespace, like `42 .bar`. And, combining that without a space is
a syntax error (presumably due to ambiguity between fractions and
dot notation for property).

I considered addressing this by recognising TYPE_DOTLESSNUM as
distinct from TYPE_LITERAL. We could then use this to enter new
variations of expression states, and create a list of these states
(akin to $divStates), and check that when checking $opChars to
decide on the $pad assignment.

I moved away from this because it would double the size of the
model. This is fine for performance (it's relatively small and
constant memory), but makes maintenance horrible by creating even
more copies of EXPRESSION which we already have many of (see PS1
for an unfinished version of that).

Instead, I'm adding a lookbehind to decide on it retroactively.
This feels dirty and won't scale, but might work for now.

Bug: T303827
Change-Id: I4b89454a5cddd97466b98d7b7f5dc102d0207e54