JavaScriptMinifier: Add basic ES2017 async-await syntax support
Code using this, could previously lead to a syntax error after
minification in rare cases where code ommitted semicolons, made use
of async as literals, or where the 1000-char linebreak would appear
between 'async' and 'function'.

This was because of the token before the function, since it's
a TYPE_LITERAL before the function, the normal behavior for all
TYPE_LITERAL in an expression is to set the next state to EXPRESSION_OP.
To fix this we first recognise `async` and `await` as their own tokens
based on ECMAScript spec, and then implement enough of their state
changes to ensure the right behaviour.

This add support for:
* async function declarations (statement)
* async function expressions (e.g. in variable assignment),
* async function in parenthetical expression (e.g. function arguments,
  or as array literal value).
* await keyword.

It does not yet handle async methods in object literals.

Bug: T343499
Change-Id: I9a3972691b36a77cc2d86d1179bed62f9239fcfa