build: Update Phan, disable sec plugin until PHP 8 compatible
This repo doesn't benefit from the sec plugin as nothing here outputs,
stores, saves, or claims to sanitize/untaint anything. If hypothetically
it was misused for that purpose in MediaWiki, Phan would catch it there.

Without this change, `composer test` fails by default for devs locally
which is confusing/unhelpful to the team.

Change-Id: I69a2c0c5801013716081ce64b2404b87b35c8203