JavaScriptMinifier: Change minify() to no longer return false
JavaScriptMinifier is a loose and tolerant parser that only strips
comments and whitespace. It already by does not aim to perform syntax
validation, and indeed for the vast majority of potentially invalid
input, the "state machine" mapping in JavaScriptMinifier simply
continues to the next token, leaving the unknown token as-is.

Callers of wikimedia/minify do not expect minify() to return false,
and indeed do not expect it to turn invalid code into valid code.

Yet..., in a handful of cases where logic was expressed procedurally
(instead of through the state machine), JavaScriptMinifier did reject
input that it happened to know was invalid.

But, just because we can, doesn't mean we should, and it makes
downstream harder to reason about as it would mean every call
technically has to deal with a boolean false return.

== parserError ==

Instead, always return a string and leave these tokens as-is, like we
already do in other cases. And instead of returning boolean false
with no details, actually fix the TODO in parseError() that has been
there for years, and expose the errors to the caller via a callback.

I considered exposing it via a by-ref output parameter, but this
becomes inpractical when using the minifier non-statically via
JavaScriptMinifierState. For example, in MediaWiki the MinifierState
is constructed ResourceLoader::getOneModuleResponse(), which is also
responsible for calling getMinifiedOutput() and possibly handling
errors. However, this method passes down the state and lets
multiple chunks get added via addSourceFile() from different callers
that shouldn't be responsible for error handling. Exposing a callback
lets the caller pre-seed this responsibility and then pass it to
other classes.

I considered emitting it unconditionally, but this would produce
fairly large numbers of errors with inputs like `1.4EEEEEEE`, and
in turn make the caller responsible for deciding when to care and when
not. To mitigate this, we only emit one call per minifier run.
Realistically any consumer parsing our payload would stop after the
first error, and thus the meaning of subsequent errors is
questionable anyway.

This repo supports PHP 7.2, hence using !$error instead of ??=.

== Background ==

In T357477, we find that after switching from the JSMinPlus validator
to Peast, some cases of invalid input were now making it past the
validator and thus to the minifier. This exposed a bug where
ResourceLoader would pass a string to the minifier, and then get
boolean false back, which has never happened before.

Going forward, JavaScriptMinifier will return the string as-is, thus
allowing web responses to continue to succeed (instead of causing
HTTP 500 errors), and then in the background we now emit a message to
the error log (e.g. Logstash) where devs can notice it and address it
higher up in the stack (e.g. the validator).

Bug: T357477
Change-Id: Ib574851bbede70ddb5be5618d0b77fcf76cd7f03