CSSMin: Avoid backtracking in regex patterns where it's not needed
This is motivated by the recent discussion in T225730#10390381.
Unfortunately this optimization will not have much of an effect,
according to my quick tests. Still I think it's worth it. I can
see a small improvement when I run the tests for this lib locally.
And there is no way this can have a negative effect. The tests
prove that everything still works as it should.

Warning, don't add such extra + everywhere. There are actually
places where a regex pattern is designed to rely on backtracking,
as we can see in the first "unquoted" part of the URL regex.
I added it only where I'm sure backtracking is not needed.

Change-Id: Ia8c3054583cb337b5c026c4e71b8a4a0a0f5684d